refactor(seed-018): move export utils under shared/ + de-dup candidate_grid triage glyphs
SEED-018 non-core cleanup (zero behavior change):

#44/M5 — relocate shared_export_utils.py implementation to shared/export_utils.py
and turn the root module into a thin re-export shim. ~12 callers keep working via
`import shared_export_utils` / `from shared_export_utils import X` unchanged.
build_app.bat and GenizahSearchPro.spec add an explicit shared/export_utils.py
data entry (the shared/ dir was already bundled wholesale). New
tests/test_export_utils_shim.py pins the shim ≡ shared.export_utils public surface.

#33 (web part) — candidate_grid.py table triage glyph map de-duplicated: the
hardcoded {"yes": "✓", "maybe": "?", "no": "✗"} now derives from the shared
shared.joins_lab.TRIAGE_ICONS source of truth (matching the existing
_TRIAGE_COLORS derivation). Desktop _TRIAGE_GLYPH and correction-status dicts
untouched (separate core slice / different concept).
